pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oxidizer-lite"
version = "0.1.0"
requires-python = ">=3.10"
description = "A distributed, event-driven DAG orchestration engine built on Redis Streams"
readme = "README.md"
license = { text = "MIT" }
keywords = ["pipeline", "dag", "distributed", "redis", "data-engineering", "orchestration", "etl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "duckdb>=1.5.0",
    "redis>=5.0.0",
    "PyYAML>=6.0",
    "structlog>=24.0",
    "requests>=2.31.0",
    "boto3>=1.34.0",
    "fastmcp>=2.0",
    "memory-profiler>=0.61.0",
    "starlette>=0.37.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/moos-engineering/oxidizer-lite"
Repository = "https://github.com/moos-engineering/oxidizer-lite"
Documentation = "https://github.com/moos-engineering/oxidizer-lite#readme"
Issues = "https://github.com/moos-engineering/oxidizer-lite/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["oxidizer_lite*"]

[tool.setuptools.package-data]
oxidizer_lite = ["*.html", "*.yml"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require Docker)",
    "chaos: marks tests as chaos/failure scenario tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
